[SDX-1041] Create box component - moved BaseDropdown to general base package
