+++ src/framework/_index.scss
@@ -6,7 +6,7 @@
 @import 'global/index';
 
 // Base components
-@import 'baseDropdown/index';
+@import 'base/dropdown/index';
 
 // Components
 @import 'addOnControl/index';
