+++ src/framework/addOnControl/dropdown/AddOnDropdown.jsx
@@ -5,7 +5,7 @@ import React, {
 } from 'react';
 import classNames from 'classnames';
 
-import BaseDropdown from '../../baseDropdown/BaseDropdown.jsx';
+import BaseDropdown from '../../base/dropdown/BaseDropdown.jsx';
 
 import AddOnDropdownOption from './AddOnDropdownOption.jsx';
