+++ src/framework/dropdown/Dropdown.spec.jsx
@@ -1,7 +1,7 @@
 
 import { TestCaseFactory } from 'react-test-kit';
 import Dropdown from './Dropdown.jsx';
-import BaseDropdown from '../baseDropdown/BaseDropdown.jsx';
+import BaseDropdown from '../base/dropdown/BaseDropdown.jsx';
 
 describe('Dropdown', () => {
   const requiredProps = {
