+++ src/framework/statusDropdown/StatusDropdownOption.spec.jsx
@@ -1,7 +1,7 @@
 
 import { TestCaseFactory } from 'react-test-kit';
 import StatusDropdownOption from './StatusDropdownOption.jsx';
-import BaseDropdownOption from '../baseDropdown/BaseDropdownOption.jsx';
+import BaseDropdownOption from '../base/dropdown/BaseDropdownOption.jsx';
 
 describe('StatusDropdownOption', () => {
   describe('DOM structure', () => {
